# proxier/ipvs: Retrieve all IPVS services in a single call