**What type of PR is this?**

/kind bug

**What this PR does / why we need it**:

This PR allows kube-proxy to retrieve all IPVS services in a single netlink call to the kernel. Previously, at least two `ipvs.GetService` calls would be issued for each ClusterIP service, more for other types of service. In environments with a lot of services this would mean a huge number of context switches and a high CPU usage for kube-proxy during sync.

This PR leverages the fact that it's possible to get all ipvs services from the system in a single `ipvs.GetServices` call. This
function was already being used in the `syncProxyRules` code to cleanup old services, this PR simply moves it to an earlier point and ensures we keep a map of cached services updated with the ones the sync process needs to create.

**Special notes for your reviewer**:

I found this problem while profiling a real node with a large number of NodePort services that was continuously syncing due to some flaky endpoints in the cluster, later I wrote a potentially dangerous benchmark that is **not** included in this PR but is available in https://gist.github.com/cezarsa/09237d9500865441bbbd166d78f6973b.

The benchmark code in this gist can be executed by appending its content to `pkg/proxy/ipvs/proxier_test.go` and running `sudo go test -run xxx -bench . -benchmem`. Please note that running it will clear the local ipvs table and insert some bogus services and destinations.

These are the results of running this benchmark against an unmodified tree vs this PR:

```
name         old time/op    new time/op    delta
ClusterIP-8    42.8ms ± 0%    21.7ms ± 0%  -49.24%  (p=0.008 n=5+5)

name         old alloc/op   new alloc/op   delta
ClusterIP-8    18.1MB ± 0%     8.9MB ± 0%  -50.79%  (p=0.008 n=5+5)

name         old allocs/op  new allocs/op  delta
ClusterIP-8      243k ± 0%      119k ± 0%  -51.12%  (p=0.008 n=5+5)
```

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".

For more information on release notes see: https://git.k8s.io/community/contributors/guide/release-notes.md
-->
```release-note
Improved performance for kube-proxy in IPVS mode.
```