@andrewsykim 2020-10-27
@cezarsa what version of kube-proxy did you run your benchmarks against? We've had other performance improvements added to kube-proxy over the last few releases (e.g. https://github.com/kubernetes/kubernetes/pull/92609) and want to know if those are already accounted for in your tests.