@cezarsa 2020-10-27
@andrewsykim Hi! The benchmark was executed against the current HEAD when I created this PR (c492e5bfb78b77139651d7deac4c80d06edbeae7) so it already included the changes from #92609.