@cezarsa 2020-10-27
Yeah, that's right. To be more clear the `old time` is on master (c492e5b) adding only the benchmark code from the linked gist, and the `new time` includes commit 68d28b840207008c87f934fc905ff1b2df2d3769 from this PR.