@andrewsykim 2020-10-27
Great! Thanks for clarifying! Will review this in the next few days.

I also saw your thread on lvs-devel about this, will read into that as well for more context.