@cezarsa 2020-10-27
Oh cool that you saw that thread on lvs-devel. :) I can expand a bit more on the context of that thread related to this PR. This PR allows kube-proxy to replace many (N for each kubernetes service) IPVS_CMD_GET_SERVICE "get" calls with a single IPVS_CMD_GET_SERVICE "dump" call. However, even with this PR applied kube-proxy still needs to make many IPVS_CMD_GET_DEST "dump" calls (`ipvs.GetDestinations`), for each service, because there's currently no way to dump every service **and** their destinations from the kernel in a single call.

My proposal on lvs-devel was to expand the netlink interface to include a new command to dump every service and destinations in a single call. I'm working on this patch right now and if I eventually get it right (and if it's eventually accepted into the kernel :)) I would be able to make a new PR to kube-proxy replacing all the existing IPVS_CMD_GET_DEST "dump" calls with a single one. Honestly I'm excited by trying to mess with the kernel in some useful way and I'm treating this as a learning experience but I'm not 100% sure it should be merged.