@rikatz 2020-10-27
So as spoken with @andrewsykim I've made some tests.

From the side of scalability, in an environment with 5k services we got some improvement around 25 to 30% of CPU usage (on a machine inside the cluster, with kube-proxy enabled but without any Pod/load)

Also, per pprof we can see there's a significant reduction in the syncService function call (which makes a lot of netlink calls).

Nice work @cezarsa