@rikatz 2020-11-12
/cc
Will test tomorrow (friday) and report back